List algorithms

List algorithms

The Scheduling.jl package provides a set of algorithms based on well-known priority rules. These algorithms are designed for scheduling on identical parallel machines.

Scheduling.Algorithms.list — Method.

list(J::Vector{Job}, M::Vector{Machine}; copy = false)

Schedules jobs in the order of their appearance in the J vector. If more than one machine can be selected, it selects the machine which is first in the M vector. If copy is set to true, then the returned structure will refer to the copies of the input vectors.

source

Scheduling.Algorithms.spt — Method.

spt(J::Vector{Job}, M::Vector{Machine}; weighted = false, copy = false)

Schedules jobs in the order of their processing times, starting with the shortest one. If weighted is set to true, then the sorting will be weighted. If copy is set to true, then the returned structure will refer to the copies of the input vectors.

source

Scheduling.Algorithms.wspt — Method.

wspt(J::Vector{Job}, M::Vector{Machine}; copy = false)

Schedules jobs in the order of their processing times to weight ratios, starting with the lowest one. If copy is set to true, then the returned structure will refer to the copies of the input vectors.

source

Scheduling.Algorithms.lpt — Method.

lpt(J::Vector{Job}, M::Vector{Machine}; weighted = false, copy = false)

Schedules jobs in the order of their processing times, starting with the largest one. If weighted is set to true, then the sorting will be weighted. If copy is set to true, then the returned structure will refer to the copies of the input vectors.

source

Scheduling.Algorithms.wlpt — Method.

wlpt(J::Vector{Job}, M::Vector{Machine}; copy = false)

Schedules jobs in the order of their processing times to weight ratios, starting with the highest one. If copy is set to true, then the returned structure will refer to the copies of the input vectors.

source