Objective functions

Objective functions

Given a schedule, one can find its objective value based on one of the following functions from the Scheduling.Objectives submodule. After the module is loaded, the function names are in the global scope.

Scheduling.Objectives.cmax — Method.

cmax(S::Schedule)

Returns the maximum completion time in schedule S.

source

Scheduling.Objectives.csum — Method.

csum(S::Schedule; weighted = false)

Returns the total completion time of these jobs in schedule S which are referenced in the S.jobs vector. If weighted is set to true, then the result will be weighted.

source

Scheduling.Objectives.wcsum — Method.

wcsum(S::Schedule)

Returns the total weighted completion time of these jobs in schedule S which are referenced in the S.jobs vector. It is an alias for csum(S::Schedule, weighted = true).

source

Scheduling.Objectives.lmax — Method.

lmax(S::Schedule)

Returns the maximum lateness in schedule S. If the schedule is empty, then the function returns -Inf.

source

Scheduling.Objectives.nt — Method.

nt(S::Schedule; weighted = false)

Returns the number of tardy jobs in schedule S which are referenced in the S.jobs vector. If weighted is set to true, then the result will be weighted.

source

Scheduling.Objectives.wnt — Method.

wnt(S::Schedule)

Returns the number of tardy jobs in schedule S which are referenced in the S.jobs vector. It is an alias for nt(S::Schedule, weighted = true).

source

Scheduling.Objectives.tsum — Method.

tsum(S::Schedule; weighted = false)

Returns the total tardiness of these jobs in schedule S which are referenced in the S.jobs vector. If weighted is set to true, then the result will be weighted.

source

Scheduling.Objectives.wtsum — Method.

wtsum(S::Schedule)

Returns the total tardiness of these jobs in schedule S which are referenced in the S.jobs vector. It is an alias for tsum(S::Schedule, weighted = true).

source